AOI_Recipe_check/requirements.txt
altgraph==0.17.2
et-xmlfile==1.1.0
future==0.18.2
importlib-metadata==4.8.2
numpy==1.21.5
openpyxl==3.0.9
pandas==1.3.5
pefile==2021.9.3
Pillow==8.4.0
pyinstaller==4.7
pyinstaller-hooks-contrib==2021.4
python-dateutil==2.8.2
pytz==2021.3
pywin32-ctypes==0.2.0
six==1.16.0
typing_extensions==4.0.1
zipp==3.6.0
